Fix MEDIUM issues: resize cast, AHCI IS clear, RTC NMI mask
M2: WM resize width/height cast uses i64 as u64 for sign extension.
M5: AHCI IS register clears only PORT_IS_TFES, not all pending bits.
M6: RTC register read sets bit 7 to keep NMI disabled during access.

+++ programsv2/edos-wm/src/main.rs
@@ -251,8 +251,8 @@ fn main() {
                 // Apply the new dimensions
                 let _ = window_set(resize.window_id, property::X, new_x as i64 as u64);
                 let _ = window_set(resize.window_id, property::Y, new_y as i64 as u64);
-                let _ = window_set(resize.window_id, property::WIDTH, new_w as u64);
-                let _ = window_set(resize.window_id, property::HEIGHT, new_h as u64);
+                let _ = window_set(resize.window_id, property::WIDTH, new_w as i64 as u64);
+                let _ = window_set(resize.window_id, property::HEIGHT, new_h as i64 as u64);
             } else {
                 // Mouse released - stop resizing
                 // Send resize event to the window with final dimensions
